Extract BlockRequirement.IsSatisfied, satisfy both
Previous logic was checking either satisfied, when it really needs to satisfy both to be included in the slot list.

I extracted out some of the enums to separate files to reduce the noise, and added a raw file dumper for the narc (safariBase).
The only functional change is in `IsPermutationValid`, which now passes the span to the block threshold object for satisfy checks.

https://github.com/kwsch/PKHeX/issues/4416